Make notes easier to understand
